refactor: cleans and makes everything ready for part 2

+++ src/main/java/fi/tuomax/adventofcode/year2022/day20/GpsNumber.java
@@ -3,7 +3,7 @@ package fi.tuomax.adventofcode.year2022.day20;
 public class GpsNumber 
 {
 
-    public int num;
+    public long num;
     
     public int origIdx;
